LayoutTests/imported/w3c/web-platform-tests/content-security-policy/webrtc/webrtc-blocked-explicit-expected.txt


FAIL webrtc blocked with an explicit webrtc blocked policy assert_equals: expected "blocked" but got "allowed"
